reorganize code-with-us opportunity back-end and shared modules for consistency
