+++ src/shared/lib/resources/opportunity/code-with-us.ts
